requirements_ml.txt
opencv-python
matplotlib>=3.0.0
numpy>=1.20.0
pandas>=1.0.0
openpyxl
Pillow>=8.3.1
scipy>=1.5.0
scikit-image>=0.18.3
gdown
tensorflow
scikit-learn
joblib